backend/requirements.txt
Flask==2.3.3
mysql-connector-python==9.5.0
python-dotenv==1.2.1